requirements.txt
huggingface_hub>=0.21.0
transformers>=4.38.0
torch>=2.1.0
datasets>=2.17.0
pytest>=8.0.0
pytest-mock>=3.12.0
numpy>=1.26.0
tqdm>=4.66.0
python-dotenv>=1.0.0
psutil>=5.9.0
accelerate>=0.27.0
safetensors>=0.4.2
rich>=13.0.0
